[CI] Reorganize entrypoints CI (#44947)
