+++ tests/entrypoints/unit_tests/test_chat_utils.py
