+++ tests/entrypoints/unit_tests/test_launch_cli.py
