+++ tests/plugins_tests/lora_resolvers/test_hf_hub_resolver.py
